renamed node, clang formatted, updated README
