+++ ublox_dgnss/CMakeLists.txt
@@ -6,6 +6,7 @@ find_package(ament_cmake REQUIRED)
 find_package(ublox_ubx_msgs REQUIRED)
 find_package(ublox_ubx_interfaces REQUIRED)
 find_package(ublox_dgnss_node REQUIRED)
+find_package(ublox_nav_sat_fix_hp_node REQUIRED)
 
 install(
   DIRECTORY launch
