+++ ublox_dgnss/package.xml
@@ -13,6 +13,7 @@
   <depend>ublox_ubx_msgs</depend>
   <depend>ublox_ubx_interfaces</depend>
   <depend>ublox_dgnss_node</depend>
+  <depend>ublox_nav_sat_fix_hp_node</depend>
 
   <export>
     <build_type>ament_cmake</build_type>
